Split version controller out into many files
