+++ packages/microapps-deployer/src/controllers/version/DeployVersion.spec.ts
@@ -1,8 +1,8 @@
 /// <reference types="jest" />
 import 'reflect-metadata';
 import 'jest-dynalite/withDb';
-import { Config, IConfig } from '../config/Config';
-jest.mock('../config/Config');
+import { Config, IConfig } from '../../config/Config';
+jest.mock('../../config/Config');
 type Writeable<T> = { -readonly [P in keyof T]: T[P] };
 const theConfig: Writeable<IConfig> = {
   awsAccountID: '00123456789',
@@ -20,6 +20,7 @@ const theConfig: Writeable<IConfig> = {
   uploadRoleName: 'microapps-upload-test-role',
   rootPathPrefix: 'dev',
   requireIAMAuthorization: true,
+  parentDeployerLambdaARN: '',
 };
 const origConfig = { ...theConfig };
 Object.defineProperty(Config, 'instance', {
@@ -33,24 +34,15 @@ import * as apigwy from '@aws-sdk/client-apigatewayv2';
 import * as dynamodb from '@aws-sdk/client-dynamodb';
 import * as lambda from '@aws-sdk/client-lambda';
 import * as s3 from '@aws-sdk/client-s3';
-import * as sts from '@aws-sdk/client-sts';
 
-import {
-  IDeployVersionPreflightRequest,
-  ICreateApplicationRequest,
-  IDeployVersionRequest,
-  IDeleteVersionRequest,
-  ILambdaAliasRequest,
-  ILambdaAliasResponse,
-} from '@pwrdrvr/microapps-deployer-lib';
+import { ICreateApplicationRequest, IDeployVersionRequest } from '@pwrdrvr/microapps-deployer-lib';
 import { DBManager, Version } from '@pwrdrvr/microapps-datalib';
 import type * as lambdaTypes from 'aws-lambda';
 import { mockClient, AwsClientStub } from 'aws-sdk-client-mock';
 import sinon from 'sinon';
-import { handler, overrideDBManager } from '../index';
+import { handler, overrideDBManager } from '../../index';
 
 let s3Client: AwsClientStub<s3.S3Client>;
-let stsClient: AwsClientStub<sts.STSClient>;
 let apigwyClient: AwsClientStub<apigwy.ApiGatewayV2Client>;
 let lambdaClient: AwsClientStub<lambda.LambdaClient>;
 let dynamoClient: dynamodb.DynamoDBClient;
@@ -98,7 +90,6 @@ describe('VersionController', () => {
 
     sandbox = sinon.createSandbox();
     s3Client = mockClient(s3.S3Client).onAnyCommand().rejects();
-    stsClient = mockClient(sts.STSClient).onAnyCommand().rejects();
     apigwyClient = mockClient(apigwy.ApiGatewayV2Client).onAnyCommand().rejects();
     lambdaClient = mockClient(lambda.LambdaClient).onAnyCommand().rejects();
   });
@@ -108,468 +99,6 @@ describe('VersionController', () => {
     s3Client.restore();
     apigwyClient.restore();
     lambdaClient.restore();
-  });
-
-  describe('deleteVersion', () => {
-    it('should 404 for version that does not exist', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-
-      lambdaClient.onAnyCommand().rejects();
-      apigwyClient.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-
-      const response = await handler(
-        {
-          appName,
-          semVer,
-          type: 'deleteVersion',
-        } as IDeleteVersionRequest,
-        { awsRequestId: '123' } as lambdaTypes.Context,
-      );
-      expect(response.statusCode).toEqual(404);
-      expect(lambdaClient.calls()).toHaveLength(0);
-      expect(apigwyClient.calls()).toHaveLength(0);
-      expect(stsClient.calls()).toHaveLength(0);
-    });
-
-    it('should 200 for version that exists', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeIntegrationID = 'integration123';
-      const fakeRoute1ID = 'route123';
-      const fakeRoute2ID = 'route456';
-
-      lambdaClient
-        .on(lambda.GetAliasCommand, {
-          FunctionName: 'arn:aws:lambda:us-east-1:123456789012:function:my-function',
-          Name: 'my-alias',
-        })
-        .resolves({
-          FunctionVersion: '1',
-        })
-        .on(lambda.DeleteAliasCommand, {
-          FunctionName: 'arn:aws:lambda:us-east-1:123456789012:function:my-function',
-          Name: 'my-alias',
-        })
-        .resolves({})
-        .on(lambda.DeleteFunctionCommand, {
-          FunctionName: 'arn:aws:lambda:us-east-1:123456789012:function:my-function',
-          Qualifier: '1',
-        })
-        .resolves({})
-        .onAnyCommand()
-        .rejects();
-      s3Client
-        .onAnyCommand()
-        .rejects()
-        // Mock S3 get for staging bucket - return one file name
-        .on(s3.ListObjectsV2Command, {
-          Bucket: config.filestore.destinationBucket,
-          Prefix: `${pathPrefix}${appName}/${semVer}/`,
-        })
-        .resolves({
-          IsTruncated: true,
-          NextContinuationToken: 'nothing-to-see-here-yet',
-        })
-        .on(s3.ListObjectsV2Command, {
-          ContinuationToken: 'nothing-to-see-here-yet',
-          Bucket: config.filestore.destinationBucket,
-          Prefix: `${pathPrefix}${appName}/${semVer}/`,
-        })
-        .resolves({
-          IsTruncated: false,
-          Contents: [{ Key: `${pathPrefix}${appName}/${semVer}/index.html` }],
-        })
-        .on(s3.DeleteObjectsCommand, {
-          Bucket: config.filestore.destinationBucket,
-          Delete: {
-            Objects: [
-              {
-                Key: `${pathPrefix}${appName}/${semVer}/index.html`,
-              },
-            ],
-          },
-        })
-        .resolves({});
-      stsClient.onAnyCommand().rejects();
-      apigwyClient
-        .onAnyCommand()
-        .rejects()
-        .on(apigwy.DeleteIntegrationCommand, {
-          ApiId: config.apigwy.apiId,
-          IntegrationId: fakeIntegrationID,
-        })
-        .resolves({})
-        .on(apigwy.DeleteRouteCommand, {
-          ApiId: config.apigwy.apiId,
-          RouteId: fakeRoute1ID,
-        })
-        .resolves({})
-        .on(apigwy.DeleteRouteCommand, {
-          ApiId: config.apigwy.apiId,
-          RouteId: fakeRoute2ID,
-        })
-        .resolves({});
-
-      const version = new Version({
-        AppName: appName,
-        SemVer: semVer,
-        DefaultFile: '',
-        IntegrationID: fakeIntegrationID,
-        RouteIDAppVersion: fakeRoute1ID,
-        RouteIDAppVersionSplat: fakeRoute2ID,
-        // Note: Pending is reported as "does not exist"
-        // So don't set this to pending or the test will fail
-        Status: 'routed',
-        Type: 'lambda',
-        LambdaARN: 'arn:aws:lambda:us-east-1:123456789012:function:my-function:my-alias',
-      });
-      await version.Save(dbManager);
-
-      const request: IDeleteVersionRequest = {
-        appName,
-        semVer,
-        type: 'deleteVersion',
-      };
-
-      const response = await handler(request, { awsRequestId: '123' } as lambdaTypes.Context);
-      expect(response).toBeDefined();
-      expect(response.statusCode).toBeDefined();
-      expect(response.statusCode).toEqual(200);
-      expect(lambdaClient.calls().length).toEqual(3);
-      expect(apigwyClient.calls().length).toEqual(3);
-      expect(s3Client.calls().length).toEqual(3);
-
-      // Confirm the version record is deleted
-      const noRecord = await Version.LoadVersion({
-        dbManager,
-        key: {
-          AppName: appName,
-          SemVer: semVer,
-        },
-      });
-      expect(noRecord).toBeUndefined();
-    });
-  });
-
-  describe('deployVersionPreflight', () => {
-    it('should 404 for version that does not exist', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-
-      stsClient.on(sts.AssumeRoleCommand).resolves({
-        Credentials: {
-          AccessKeyId: 'cat',
-          SecretAccessKey: 'dog',
-          SessionToken: 'frog',
-          Expiration: new Date(),
-        },
-      });
-
-      const response = await handler(
-        {
-          appName,
-          semVer,
-          type: 'deployVersionPreflight',
-        } as IDeployVersionPreflightRequest,
-        { awsRequestId: '123' } as lambdaTypes.Context,
-      );
-      expect(response.statusCode).toEqual(404);
-    });
-
-    it('should 200 for version that exists', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-
-      stsClient.on(sts.AssumeRoleCommand).resolves({
-        Credentials: {
-          AccessKeyId: 'cat',
-          SecretAccessKey: 'dog',
-          SessionToken: 'frog',
-          Expiration: new Date(),
-        },
-      });
-
-      const version = new Version({
-        AppName: appName,
-        DefaultFile: '',
-        IntegrationID: '',
-        SemVer: semVer,
-        // Note: Pending is reported as "does not exist"
-        // So don't set this to pending or the test will fail
-        Status: 'integrated',
-        Type: 'lambda',
-      });
-      await version.Save(dbManager);
-
-      const response = await handler(
-        {
-          appName,
-          semVer,
-          type: 'deployVersionPreflight',
-        } as IDeployVersionPreflightRequest,
-        { awsRequestId: '123' } as lambdaTypes.Context,
-      );
-      expect(response.statusCode).toEqual(200);
-    });
-  });
-
-  describe('lambdaAlias', () => {
-    const fakeLambdaARNBase = `arn:aws:lambda:${config.awsRegion}:${config.awsAccountID}:function:new-app-function`;
-
-    it('should 200 for version that exists when !overwrite', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeLambdaVersion = '31';
-      const fakeLambdaARNWithVersion = `${fakeLambdaARNBase}:${fakeLambdaVersion}`;
-      const fakeLambdaAlias = 'v0_0_0';
-
-      s3Client.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-      lambdaClient.onAnyCommand().rejects();
-
-      const version = new Version({
-        AppName: appName,
-        SemVer: semVer,
-        // Note: Pending is reported as "does not exist"
-        // So don't set this to pending or the test will fail
-        Status: 'routed',
-        Type: 'lambda',
-        LambdaARN: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-      });
-      await version.Save(dbManager);
-
-      const request: ILambdaAliasRequest = {
-        appName,
-        semVer,
-        lambdaARN: fakeLambdaARNWithVersion,
-        type: 'lambdaAlias',
-      };
-      const response = (await handler(request, {
-        awsRequestId: '123',
-      } as lambdaTypes.Context)) as ILambdaAliasResponse;
-
-      expect(response.statusCode).toBe(200);
-      expect(response.type).toBe('lambdaAlias');
-      expect(response.lambdaAliasARN).toBe(`${fakeLambdaARNBase}:${fakeLambdaAlias}`);
-      expect(lambdaClient.calls()).toHaveLength(0);
-    });
-
-    it('should 200 for version that exists when overwrite=true', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeLambdaVersion = '31';
-      const fakeLambdaARNWithVersion = `${fakeLambdaARNBase}:${fakeLambdaVersion}`;
-      const fakeLambdaAlias = 'v0_0_0';
-
-      s3Client.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-      lambdaClient
-        .onAnyCommand()
-        .rejects()
-        .on(lambda.GetAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-        })
-        .resolves({
-          AliasArn: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-          FunctionVersion: fakeLambdaVersion,
-        });
-
-      const version = new Version({
-        AppName: appName,
-        SemVer: semVer,
-        // Note: Pending is reported as "does not exist"
-        // So don't set this to pending or the test will fail
-        Status: 'routed',
-        Type: 'lambda',
-        LambdaARN: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-      });
-      await version.Save(dbManager);
-
-      const request: ILambdaAliasRequest = {
-        appName,
-        semVer,
-        lambdaARN: fakeLambdaARNWithVersion,
-        type: 'lambdaAlias',
-        overwrite: true,
-      };
-      const response = (await handler(request, {
-        awsRequestId: '123',
-      } as lambdaTypes.Context)) as ILambdaAliasResponse;
-
-      expect(response.statusCode).toBe(200);
-      expect(response.actionTaken).toBe('verified');
-      expect(response.type).toBe('lambdaAlias');
-      expect(response.lambdaAliasARN).toBe(`${fakeLambdaARNBase}:${fakeLambdaAlias}`);
-      expect(lambdaClient.calls()).toHaveLength(1);
-    });
-
-    it('should 201 for version that !exists when !overwrite', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeLambdaVersion = '31';
-      const fakeLambdaARNWithVersion = `${fakeLambdaARNBase}:${fakeLambdaVersion}`;
-      const fakeLambdaAlias = 'v0_0_0';
-
-      s3Client.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-      lambdaClient
-        .onAnyCommand()
-        .rejects()
-        .on(lambda.GetAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-        })
-        .rejects({
-          name: 'ResourceNotFoundException',
-        })
-        .on(lambda.CreateAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-          FunctionVersion: fakeLambdaVersion,
-        })
-        .resolves({
-          AliasArn: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-          FunctionVersion: fakeLambdaVersion,
-        });
-
-      const request: ILambdaAliasRequest = {
-        appName,
-        semVer,
-        lambdaARN: fakeLambdaARNWithVersion,
-        type: 'lambdaAlias',
-      };
-      const response = (await handler(request, {
-        awsRequestId: '123',
-      } as lambdaTypes.Context)) as ILambdaAliasResponse;
-
-      expect(response.statusCode).toBe(201);
-      expect(response.type).toBe('lambdaAlias');
-      expect(response.lambdaAliasARN).toBe(`${fakeLambdaARNBase}:${fakeLambdaAlias}`);
-      expect(lambdaClient.calls()).toHaveLength(2);
-    });
-
-    it('should 200 for version that exists when overwrite and version changed', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeLambdaVersionStart = '3';
-      const fakeLambdaVersionEnd = '31';
-      const fakeLambdaARNWithVersionEnd = `${fakeLambdaARNBase}:${fakeLambdaVersionEnd}`;
-      const fakeLambdaAlias = 'v0_0_0';
-
-      s3Client.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-      lambdaClient
-        .onAnyCommand()
-        .rejects()
-        .on(lambda.GetAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-        })
-        .resolves({
-          AliasArn: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-          FunctionVersion: fakeLambdaVersionStart,
-        })
-        .on(lambda.UpdateAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-          FunctionVersion: fakeLambdaVersionEnd,
-        })
-        .resolves({
-          AliasArn: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-          FunctionVersion: fakeLambdaVersionEnd,
-        });
-
-      const version = new Version({
-        AppName: appName,
-        SemVer: semVer,
-        // Note: Pending is reported as "does not exist"
-        // So don't set this to pending or the test will fail
-        Status: 'routed',
-        Type: 'lambda',
-        LambdaARN: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-      });
-      await version.Save(dbManager);
-
-      const request: ILambdaAliasRequest = {
-        appName,
-        semVer,
-        lambdaARN: fakeLambdaARNWithVersionEnd,
-        type: 'lambdaAlias',
-        overwrite: true,
-      };
-      const response = (await handler(request, {
-        awsRequestId: '123',
-      } as lambdaTypes.Context)) as ILambdaAliasResponse;
-
-      expect(response.statusCode).toBe(200);
-      expect(response.actionTaken).toBe('updated');
-      expect(response.type).toBe('lambdaAlias');
-      expect(response.lambdaAliasARN).toBe(`${fakeLambdaARNBase}:${fakeLambdaAlias}`);
-      expect(lambdaClient.calls()).toHaveLength(2);
-    });
-
-    it('should 201 for function ARN that !exists when !overwrite', async () => {
-      const appName = 'newapp';
-      const semVer = '0.0.0';
-      const fakeLambdaVersion = '31';
-      const fakeLambdaAlias = 'v0_0_0';
-
-      s3Client.onAnyCommand().rejects();
-      stsClient.onAnyCommand().rejects();
-      lambdaClient
-        .onAnyCommand()
-        .rejects()
-        .on(lambda.PublishVersionCommand, {
-          FunctionName: fakeLambdaARNBase,
-        })
-        .resolves({
-          FunctionName: fakeLambdaARNBase,
-          Version: fakeLambdaVersion,
-        })
-        .on(lambda.GetFunctionCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Qualifier: fakeLambdaVersion,
-        })
-        .resolves({
-          Configuration: {
-            LastUpdateStatus: 'Successful',
-          },
-        })
-        .on(lambda.GetAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-        })
-        .rejects({
-          name: 'ResourceNotFoundException',
-        })
-        .on(lambda.CreateAliasCommand, {
-          FunctionName: fakeLambdaARNBase,
-          Name: fakeLambdaAlias,
-          FunctionVersion: fakeLambdaVersion,
-        })
-        .resolves({
-          AliasArn: `${fakeLambdaARNBase}:${fakeLambdaAlias}`,
-          FunctionVersion: fakeLambdaVersion,
-        });
-
-      const request: ILambdaAliasRequest = {
-        appName,
-        semVer,
-        lambdaARN: fakeLambdaARNBase,
-        type: 'lambdaAlias',
-      };
-      const response = (await handler(request, {
-        awsRequestId: '123',
-      } as lambdaTypes.Context)) as ILambdaAliasResponse;
-
-      expect(response.statusCode).toBe(201);
-      expect(response.type).toBe('lambdaAlias');
-      expect(response.lambdaAliasARN).toBe(`${fakeLambdaARNBase}:${fakeLambdaAlias}`);
-      expect(lambdaClient.calls()).toHaveLength(4);
-    });
   });
 
   describe('deployVersion - lambda (apigwy)', () => {
