+++ packages/microapps-deployer/src/controllers/version/index.ts
@@ -1,0 +1,4 @@
+export * from './DeleteVersion';
+export * from './DeployVersion';
+export * from './DeployVersionPreflight';
+export * from './LambdaAlias';
