+++ packages/microapps-deployer/src/lib/GetAppNameOrRootTrailingSlash.ts
@@ -1,0 +1,8 @@
+import { IDeployVersionRequestBase } from '@pwrdrvr/microapps-deployer-lib';
+
+export function GetAppNameOrRootTrailingSlash(
+  request: Pick<IDeployVersionRequestBase, 'appName'>,
+): string {
+  const isRootApp = request.appName === '[root]';
+  return isRootApp ? '' : `${request.appName}/`;
+}
